Data.Transient.Frozen

Documentation

class Frozen t where

Associated Types

data Transient t :: * -> k -> *

Methods

freeze :: PrimMonad m => Transient t (PrimState m) a -> m (t a)

O(1). amortized conversion from a mutable structure to an immutable one

thaw :: t a -> Transient t s a

O(1). worst-case conversion from an immutable structure to a mutable one

modify :: Frozen t => (forall s. Transient t s a -> ST s (Transient t s b)) -> t a -> t b

O(1) + modification time. to modify an immutable structure with mutable operations.

The transient passed to this function is considered frozen

query :: (PrimMonad m, Frozen t) => (t a -> r) -> Transient t (PrimState m) a -> m r

O(1) + query time. Query a mutable structure with queries designed for an immutable structure.

This does _not_ destroy the transient, although, it does mean subsequent actions need to copy-on-write from scratch.

After query f wm, wm is considered frozen and may be reused.