Data.Transient.WordMap.Internal

Description

This module suppose a Word64-based array-mapped PATRICIA Trie.

The most significant nybble is isolated by using techniques based on https://www.fpcomplete.com/user/edwardk/revisiting-matrix-multiplication/part-4 but modified to work nybble-by-nybble rather than bit-by-bit.

This structure secretly maintains a finger to the previous mutation to speed access and repeated operations.

Documentation

type Key = Word64

type Mask = Word16

type Offset = Int

Utilities

ptrEq :: a -> a -> Bool

ptrNeq :: a -> a -> Bool

index :: Word16 -> Word16 -> Int

level :: Key -> Int

Note: level 0 will return a negative shift, so don't use it

maskBit :: Key -> Offset -> Int

mask :: Key -> Offset -> Word16

apogeeBit :: Key -> Key -> Int

Given a pair of keys, they agree down to this height in the display, don't use this when they are the same

apogeeBit k ok = unsafeShiftR (level (xor k ok)) 2
level (xor k ok) = unsafeShiftL (apogeeBit k ok) 2

apogee :: Key -> Key -> Mask

WordMaps

data Node a

Constructors

Full !Key !Offset !(SmallArray (Node a))

Node !Key !Offset !Mask !(SmallArray (Node a))

Tip !Key a

data WordMap a

Constructors

WordMap

Fields

fingerKey :: !Key

fingerMask :: !Mask

fingerValue :: !(Maybe a)

fingerPath :: !(SmallArray (Node a))

data TNode s a

Constructors

TFull !Key !Offset !(SmallMutableArray s (TNode s a))

TNode !Key !Offset !Mask !(SmallMutableArray s (TNode s a))

TTip !Key a

data TWordMap s a

This is a transient WordMap with a clojure-like API

Constructors

TWordMap

Fields

transientFingerKey :: !Key

transientFingerMask :: !Mask

transientFingerValue :: !(Maybe a)

transientFingerPath :: !(SmallMutableArray s (TNode s a))

persisted :: (forall s. TWordMap s a) -> WordMap a

unsafePersistentTNode :: TNode s a -> Node a

unsafePersistent :: TWordMap s a -> WordMap a

newtype MWordMap s a

This is a mutable WordMap with a classic Haskell mutable container-style API

On the plus side, this API means you don't need to carefully avoid reusing a transient On the minus side, you have an extra reference to track.

Constructors

MWordMap

Fields

runMWordMap :: MutVar s (TWordMap s a)

thaw :: PrimMonad m => WordMap a -> m (MWordMap (PrimState m) a)

freeze :: PrimMonad m => MWordMap (PrimState m) a -> m (WordMap a)

Transient WordMaps

transient :: WordMap a -> TWordMap s a

O(1) worst-case conversion from an immutable structure to a mutable one

persistent :: PrimMonad m => TWordMap (PrimState m) a -> m (WordMap a)

O(1) amortized conversion from a mutable structure to an immutable one

empty :: WordMap a

emptySmallMutableArray :: SmallMutableArray s a

emptyT :: TWordMap s a

emptyM :: PrimMonad m => m (MWordMap (PrimState m) a)

singleton :: Key -> a -> WordMap a

Build a singleton WordMap

singletonT :: Key -> a -> TWordMap s a

singletonM :: PrimMonad m => Key -> a -> m (MWordMap (PrimState m) a)

lookup :: Key -> WordMap a -> Maybe a

lookupM :: PrimMonad m => Key -> MWordMap (PrimState m) a -> m (Maybe a)

lookupT :: PrimMonad m => Key -> TWordMap (PrimState m) a -> m (Maybe a)

lookupTNode :: Key -> TNode s a -> ST s (Maybe a)

modify :: (forall s. TWordMap s a -> ST s (TWordMap s b)) -> WordMap a -> WordMap b

Modify an immutable structure with mutable operations.

modify f wm passed f a "persisted" transient that may be reused.

modifyM :: PrimMonad m => (TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)) -> MWordMap (PrimState m) a -> m ()

Modify a mutable wordmap with a transient operation.

query :: PrimMonad m => (WordMap a -> r) -> TWordMap (PrimState m) a -> m r

Query a transient structure with queries designed for an immutable structure.

This does _not_ destroy the transient, although, it does mean subsequent actions need to copy-on-write from scratch.

After query f wm, wm is considered persisted and may be reused.

queryM :: PrimMonad m => (WordMap a -> r) -> MWordMap (PrimState m) a -> m r

Query a mutable structure with queries designed for an immutable structure.

Construction

type Hint s = forall x. SmallMutableArray# s x -> State# s -> State# s

warm :: Hint s

cold :: Hint s

apply :: PrimMonad m => Hint (PrimState m) -> SmallMutableArray (PrimState m) a -> m ()

insertSmallArray :: Hint s -> SmallMutableArray s a -> Int -> a -> ST s (SmallMutableArray s a)

deleteSmallArray :: Hint s -> SmallMutableArray s a -> Int -> ST s (SmallMutableArray s a)

node :: Key -> Offset -> Mask -> SmallMutableArray s (TNode s a) -> TNode s a

fork :: Hint s -> Key -> TNode s a -> Key -> TNode s a -> ST s (TNode s a)

unplug :: Hint s -> Key -> TNode s a -> ST s (TNode s a)

canonical :: WordMap a -> Maybe (Node a)

plug :: Hint s -> Key -> TNode s a -> TNode s a -> ST s (TNode s a)

plugPath :: Hint s -> Key -> Int -> Int -> TNode s a -> SmallMutableArray s (TNode s a) -> ST s (TNode s a)

Given k located under acc, plugPath k i t acc ns plugs acc recursively into each of the nodes of ns from [i..t-1] from the bottom up

unplugPath :: Hint s -> Key -> Int -> Int -> SmallMutableArray s (TNode s a) -> ST s (TNode s a)

replugPath :: PrimMonad m => Hint (PrimState m) -> Key -> Int -> Int -> Maybe v -> SmallMutableArray (PrimState m) (TNode (PrimState m) v) -> m (TNode (PrimState m) v)

unI# :: Int -> Int#

trimT :: PrimMonad m => TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

O(1) This function enables us to GC the items that lie on the path to the finger.

Normally we only do this lazily as the finger moves out of a given area, but if you have particularly burdensome items for the garbage collector it may be worth paying this price to explicitly allow them to go free.

trimM :: PrimMonad m => MWordMap (PrimState m) a -> m ()

O(1) This function enables us to GC the items that lie on the path to the finger.

Normally we only do this lazily as the finger moves out of a given area, but if you have particularly burdensome items for the garbage collector it may be worth paying this price. to explicitly allow them to go free.

trim :: WordMap a -> WordMap a

O(1) This function enables us to GC the items that lie on the path to the finger.

Normally we only do this lazily as the finger moves out of a given area, but if you have particularly burdensome items for the garbage collector it may be worth paying this price. to explicitly allow them to go free.

focusWithHint :: PrimMonad m => Hint (PrimState m) -> Key -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

focusT :: PrimMonad m => Key -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

This changes the location of the focus in a transient map. Operations near the focus are considerably cheaper.

focusT k wm invalidates any unpersisted transient wm it is passed

focusM :: PrimMonad m => Key -> MWordMap (PrimState m) a -> m ()

This changes the location of the focus in a mutable map. Operations near the focus are considerably cheaper.

focus :: Key -> WordMap a -> WordMap a

This changes the location of the focus in an immutable map. Operations near the focus are considerably cheaper.

insertWithHint :: PrimMonad m => Hint (PrimState m) -> Key -> a -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

insertT :: PrimMonad m => Key -> a -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

Transient insert.

insertT k v wm invalidates any unpersisted transient wm it is passed

insertM :: PrimMonad m => Key -> a -> MWordMap (PrimState m) a -> m ()

Mutable insert.

insert :: Key -> a -> WordMap a -> WordMap a

Immutable insert.

deleteWithHint :: PrimMonad m => Hint (PrimState m) -> Key -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

deleteT :: PrimMonad m => Key -> TWordMap (PrimState m) a -> m (TWordMap (PrimState m) a)

Transient delete. deleteT k v wm invalidates any unpersisted transient it is passed.

deleteM :: PrimMonad m => Key -> MWordMap (PrimState m) a -> m ()

Mutable delete.

delete :: Key -> WordMap a -> WordMap a

Immutable delete.

Instances