Data.Transient.WordMap

Description

This module suppose a Word64-based array-mapped PATRICIA Trie.

The most significant nybble is isolated by using techniques based on https://www.fpcomplete.com/user/edwardk/revisiting-matrix-multiplication/part-4 but modified to work nybble-by-nybble rather than bit-by-bit.

Documentation

data WordMap v

singleton :: Key -> v -> WordMap v

Build a singleton WordMap

empty :: WordMap a

insert :: Key -> v -> WordMap v -> WordMap v

delete :: Key -> WordMap v -> WordMap v

lookup :: Key -> WordMap v -> Maybe v

member :: Key -> WordMap v -> Bool

fromList :: IsList l => [Item l] -> l

The fromList function constructs the structure l from the given list of Item l